% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_tax.R
\name{cor_tax}
\alias{cor_tax}
\title{Correlate taxa abundances with lipid abundances}
\usage{
cor_tax(phy, lipids, thresh = 0.75, cutoff = 0.03,
  method = "pearson", level = "Phylum", exempt = FALSE)
}
\arguments{
\item{phy}{Phyloseq object containing an OTU and taxonomy table.}

\item{lipids}{Data frame with concentration of lipids (or any other variables/parameters) to correlate with phylogenetic data. Rows are observations/samples and must match the samples in 'phy'. Used by 'phlo_lipids'}

\item{thresh}{threshold correlation coefficient, above which OTUs will be considered (default = 0.75)}

\item{cutoff}{cutoff below which OTUs will be subsumed as "other" (applies for each compound/lipid separately)}

\item{method}{correaltion method for 'cor' {stats}, i.e., "pearson", "kendall", or "spearman"}

\item{level}{character string with taxonomic level of analysis, e.g., "Phylum", "Class", etc}

\item{exempt}{taxonomic unit that is expemt from the cutoff, i.e., is always displayed regardless of count numbers}
}
\value{
A data frame with each row representing one OTU, and each column one lipid compound. Values are character strings (phylogenetic affiliations) of above-threshold OTUs. 'lower' = below threshold, 'other' = sumsumed according to cutoff fraction
}
\description{
This function allows you to extract the phylogeneic affiliations of OTUs, using a theshold correlation coefficient
Returns a table with each row representing one OTU, and each column one lipid compound. Values are affiliations of above-threshold OTUs.
'lower' = below threshold, 'other' = sumsumed according to cutoff fraction
}
\seealso{
Other Main functions: \code{\link{cor_phy}},
  \code{\link{norm_dna}}, \code{\link{phylo_lipids}}
}
\concept{Main functions}
