% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.lipids.R
\docType{package}
\name{phylo.lipids-package}
\alias{phylo.lipids}
\alias{phylo.lipids-package}
\title{phylo.lipids: Functions to compare environmental lipidomic and meta taxonomic data}
\description{
Allows you to correlate ribosomal RNA gene-based meta-taxonomic data with concentrations of lipid biomarkers, 
   and to visualize the phylogenetic affiliation of corresponding taxa.
}
\details{
Main functions:

phylo_lipids(), cor_phy(), cor_tax(), norm_dna(),

Utility functions:

remove_clades(), rm_singletons(), trim_taxa_samp(), trim_taxa_sum()
}
\references{
Yuki Weber, Jaap S. Sinninghe Damsté, Jakob Zopfi, Cindy De Jonge, Adrian Gili, Carsten J. Schubert, Fabio Lepori, Moritz F. Lehmann, Helge Niemann (2018). "Redox-dependent niche differentiation of tetraether lipid producing bacteria: Evidence for multiple branched GDGT sources in lakes"
}
\seealso{
vignette("phylo.lipids")
}
\keyword{internal}
