% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_lipids.R
\name{phylo_lipids}
\alias{phylo_lipids}
\title{Phylogenetic diversity of correlated taxa}
\usage{
phylo_lipids(phy, lipids, method = c("pearson", "spearman", "kendall"),
  thresh = 0.75, level = "Phylum", cutoff = 0, exempt = FALSE,
  n.col = 3, sort.tax = c("alpha", "counts"), plot.theme = FALSE,
  text.size = 12, test = FALSE, test.size = 102, guide.col = 1)
}
\arguments{
\item{phy}{Phyloseq object containing an OTU and taxonomy table.}

\item{lipids}{Data frame with concentration of lipids (or any other variables/parameters) to correlate with phylogenetic data. Rows are observations/samples and must match the samples in 'phy'.}

\item{method}{correaltion coefficient used in 'cor' {stats}, i.e., "pearson", "kendall", or "spearman"}

\item{thresh}{threshold correlation coefficient, above which OTUs will be considered (default = 0.75)}

\item{level}{character string with taxonomic level of analysis, e.g., "Phylum", "Class", etc}

\item{cutoff}{cutoff below which OTUs will be subsumed as "other" (applies for each compound/lipid separately)}

\item{exempt}{taxonomic unit(s) that is/are expemt from the cutoff, i.e., is always displayed on top regardless of cutoff value and sorting}

\item{n.col}{number of columns for pie charts (default = 3)}

\item{sort.tax}{sort taxa aphabetically or by counts. Either "alpha" or "counts" (default = "alpha")}

\item{plot.theme}{a ggplot theme object which overrides defualt plot theme}

\item{text.size}{global text size, titles and labels are relative (default = 12)}

\item{test}{randomply subsample otus for testing large data sets (default = FALSE)}

\item{test.size}{number of random otus to draw (default = 100)}

\item{guide.col}{number of legend columns}
}
\value{
A list, containing a pie chart (.$pies), a table in long format (.$table), and a summary text (.$summary).
}
\description{
This function correlates absolute abundances of OTUs vs lipids, and summarizes the phylogenetic diversity of correlated taxa for each lipid.
}
\seealso{
vignette("phylo.lipids")

?phylo.lipids

Other Main functions: \code{\link{cor_phy}},
  \code{\link{cor_tax}}, \code{\link{norm_dna}}
}
\concept{Main functions}
\keyword{DNA}
\keyword{charts}
\keyword{correlation}
\keyword{lipids}
\keyword{pie}
