% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_clades.R
\name{remove_clades}
\alias{remove_clades}
\title{Removing taxonomic groups from Phyloseq objects}
\usage{
remove_clades(phy, clades)
}
\arguments{
\item{phy}{Phyloseq object}

\item{clades}{Character vector with taxonomic groups to be removed}
}
\description{
This function removes taxonomic groups from Phyloseq objects. Clades can be of different ranks.
}
\keyword{groups}
\keyword{remove}
\keyword{taxonomic}
