% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_singletons.R
\name{rm_singletons}
\alias{rm_singletons}
\title{Removes singletons from phyloseq object}
\usage{
rm_singletons(phy)
}
\arguments{
\item{phy}{Phyloseq object containing an OTU and taxonomy table.}
}
\description{
i.e., OTUs with only one read across all samples
}
\examples{

}
