% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_taxa_samp.R
\name{trim_taxa_samp}
\alias{trim_taxa_samp}
\title{trim by occurrence in min number/fraction of samples}
\usage{
trim_taxa_samp(phy, min, fraction = F)
}
\arguments{
\item{phy}{Phyloseq object containing an OTU and taxonomy table.}

\item{min}{threshold for taxa to return}

\item{fraction}{if TRUE, a relative threshold is used
i.e., 'min' das to be a frction between 0 and 1}
}
\description{
trim by occurrence in min number/fraction of samples
}
\examples{

}
